requirements.txt
torch>=2.0
torchvision>=0.15
numpy
Pillow
scikit-learn
tqdm
matplotlib
tensorboard
facenet-pytorch
tensorboard